# Implement data-specific cartesian index